+++ src/gui/text/qrawfont_ft.cpp
@@ -100,7 +100,7 @@ void QRawFontPrivate::platformCleanUp()
     // Font engine handles all resources
 }
 
-void QRawFontPrivate::platformLoadFromData(const QByteArray &fontData, int pixelSize,
+void QRawFontPrivate::platformLoadFromData(const QByteArray &fontData, qreal pixelSize,
                                            QFont::HintingPreference hintingPreference)
 {
     Q_ASSERT(fontEngine == 0);
